显示博客内容

+++ wanblog/app/src/main/java/com/wanblog/model/http/ApiManager.kt
@@ -106,7 +106,7 @@ object ApiManager {
     fun getBlogList(currentPage: Int, size: Int): Flowable<MyHttpResponse<MutableList<BlogBean>>> =
         mApiService.getBlogList(currentPage, size)
 
-    fun getBlogDetail(id: Int): Flowable<MyHttpResponse<BlogBean>> =
+    fun getBlogDetail(id: Long): Flowable<MyHttpResponse<BlogBean>> =
         mApiService.getBlogDetail(id)
