+++ wanblog/app/src/main/java/com/wanblog/presenter/contract/BlogContract.kt
@@ -14,7 +14,7 @@ interface BlogContract {
     }
 
     interface Presenter<in T> : BaseContract.BasePresenter<T> {
-        fun blogDetail(blogId: Int)
+        fun blogDetail(blogId: Long)
         fun editBlog(blog: BlogBean)
     }
